asyncio: syncronize I/O loops with mainloop
The read/write callback may directly close a file descriptor and then
enqueue a callback to terminate the I/O loop.

We need to make sure that the callback happens before the I/O loop
continues. Otherwise we either hit an annoying "file descriptor is closed"
error or, worse, access a different just-opened file that happened to
re-use the fileno.