Initial code to listen to SIGCHLD.
This change adds an easy-to-use async "wait_for_child(pid)" method which
returns the exit code / negative signal number of a child process, as soon
as it exits.

Since SIGCHLD is fraught with perils and there's multiple ways to handle
the trade-offs involved, there's a way to plug in a different handlers.
Several are provided.